// src/api/generated/model/teamCreate.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { TeamStatus } from "./teamStatus";

/**
 * Schema for creating a new team.

Attributes:
    name: Name of the team.
    description: Description of the team.
    logo: URL or path to the team logo.
    member_ids: List of user IDs to include in the team.
    status: Status of the team (default: DRAFT).
 */
export interface TeamCreate {
    /**
     * @minLength 1
     * @maxLength 255
     */
    name: string;
    description?: string | null;
    logo?: string | null;
    member_ids?: string[];
    leader_id?: string | null;
    status?: TeamStatus;
}