// src/api/generated/model/teamUpdate.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { TeamStatus } from "./teamStatus";

/**
 * Schema for updating an existing team.

Only allows updating basic team information.
Members management should be handled through separate endpoints.

Attributes:
    name: Updated name of the team.
    description: Updated description of the team.
    logo: Updated URL or path to the team logo.
    status: Updated status of the team in the contest.
    leader_id: Updated leader ID for the team.
 */
export interface TeamUpdate {
    name?: string | null;
    description?: string | null;
    logo?: string | null;
    status?: TeamStatus | null;
    leader_id?: string | null;
}